requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
click==8.1.2
Deprecated==1.2.13
Django==4.0.4
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
factory-boy==3.2.1
Faker==13.3.4
h11==0.13.0
Markdown==3.3.6
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.2
six==1.16.0
sqlparse==0.4.2
uvicorn==0.17.6
wrapt==1.14.0
